Finished grade 2
I think grade 2 is finished now, also started experimenting a litle bit
with some ides. Didnt put in to much effort in the modelling since the
whole thing feels like i should just stick to what it says in the
description for the workshop and not overcomplicate/develop further the
model.